Make a place for the version command
